Refactoring CIS scoring to support baseline calculation
