+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Business/Maturity/CisQuestionsBusiness.cs
@@ -44,7 +44,7 @@ namespace CSETWebCore.Business.Maturity
 
         private List<MATURITY_GROUPINGS> allGroupings;
 
-        private List<FlatQuestion> allWeights;
+       // private List<FlatQuestion> allWeights;
 
 
         /// <summary>
@@ -72,18 +72,9 @@ namespace CSETWebCore.Business.Maturity
             this._context = context;
             this._assessmentUtil = assessmentUtil;
             this._assessmentId = assessmentId;
-            allWeights = new List<FlatQuestion>();
 
 
-            var amm = _context.AVAILABLE_MATURITY_MODELS.Where(x => x.Assessment_Id == assessmentId).FirstOrDefault();
-            if (amm != null)
-            {
-                this._maturityModelId = amm.model_id;
-            }
-            else
-            {
-                //throw new Exception("CisQuestionsBusiness cannot be instantiated for an assessment without a maturity model.");
-            }
+
 
             // Get the baseline assessment if one is selected
             var info = _context.INFORMATION.Where(x => x.Id == assessmentId).FirstOrDefault();
@@ -92,329 +83,7 @@ namespace CSETWebCore.Business.Maturity
                 _baselineAssessmentId = info.Baseline_Assessment_Id;
             }
         }
-
-
-        /// <summary>
-        /// Returns the grouping/question/option structure for a section.
-        /// </summary>
-        /// <param name="sectionId"></param>
-        /// <returns></returns>
-        public CisQuestions GetSection(int sectionId)
-        {
-            int? sectionId1 = null;
-            if (sectionId != 0)
-            {
-                sectionId1 = sectionId;
-            }
-
-
-
-            LoadStructure(sectionId1);
-
-            // include score
-            if (sectionId1 != null)
-            {
-                this.QuestionsModel.GroupingScore = CalculateGroupingScore((int)sectionId1);
-            }
-
-            return this.QuestionsModel;
-        }
-
-
-        /// <summary>
-        /// Gathers questions and answers and builds them into a basic hierarchy.
-        /// </summary>
-        private void LoadStructure(int? sectionId)
-        {
-            QuestionsModel = new CisQuestions
-            {
-                AssessmentId = this._assessmentId
-            };
-
-            allQuestions = _context.MATURITY_QUESTIONS
-                .Include(x => x.Maturity_LevelNavigation)
-                .Include(x => x.MATURITY_REFERENCE_TEXT)
-                .Where(q =>
-                _maturityModelId == q.Maturity_Model_Id).ToList();
-
-            allAnswers = _context.ANSWER
-                .Where(a => a.Question_Type == Constants.Constants.QuestionTypeMaturity
-                    && a.Assessment_Id == this._assessmentId)
-                .ToList();
-
-            // include baseline answers
-            if (this._baselineAssessmentId != null)
-            {
-                baselineAllAnswers = _context.ANSWER
-                .Where(a => a.Question_Type == Constants.Constants.QuestionTypeMaturity
-                    && a.Assessment_Id == this._baselineAssessmentId)
-                .ToList();
-            }
-
-
-            // Get all subgroupings for this maturity model
-            allGroupings = _context.MATURITY_GROUPINGS
-                .Include(x => x.Type)
-                .Where(x => x.Maturity_Model_Id == _maturityModelId).ToList();
-
-            GetSubgroups(QuestionsModel, null, sectionId);
-        }
-
-
-        /// <summary>
-        /// Recursive method for traversing the structure.
-        /// If the filterId is specified, the subgroups are reduced to that one.
-        /// </summary>
-        private void GetSubgroups(object oParent, int? parentId, int? filterId = null)
-        {
-            var mySubgroups = allGroupings.Where(x => x.Parent_Id == parentId).OrderBy(x => x.Sequence).ToList();
-
-            if (filterId != null)
-            {
-                mySubgroups = allGroupings.Where(x => x.Grouping_Id == filterId).ToList();
-            }
-
-            if (mySubgroups.Count == 0)
-            {
-                return;
-            }
-
-            foreach (var sg in mySubgroups)
-            {
-                var nodeName = System.Text.RegularExpressions
-                    .Regex.Replace(sg.Type.Grouping_Type_Name, " ", "_");
-
-                var grouping = new Grouping()
-                {
-                    GroupType = nodeName,
-                    Abbreviation = sg.Abbreviation,
-                    GroupingId = sg.Grouping_Id,
-                    Title = sg.Title,
-                    Description = sg.Description
-                };
-
-                if (_includeText)
-                {
-                    grouping.Description = sg.Description;
-                }
-
-                if (oParent is CisQuestions)
-                {
-                    ((CisQuestions)oParent).Groupings.Add(grouping);
-                }
-
-                if (oParent is Grouping)
-                {
-                    ((Grouping)oParent).Groupings.Add(grouping);
-                }
-
-
-
-                // are there any questions that belong to this grouping?
-                var myQuestions = allQuestions.Where(x => x.Grouping_Id == sg.Grouping_Id
-                    && x.Parent_Question_Id == null && x.Parent_Option_Id == null).ToList();
-
-                foreach (var myQ in myQuestions.OrderBy(s => s.Sequence))
-                {
-                    var answer = allAnswers.FirstOrDefault(x => x.Question_Or_Requirement_Id == myQ.Mat_Question_Id);
-
-                    var question = new Model.Cis.Question()
-                    {
-                        QuestionId = myQ.Mat_Question_Id,
-                        Sequence = myQ.Sequence,
-                        DisplayNumber = myQ.Question_Title,
-                        ParentQuestionId = myQ.Parent_Question_Id,
-                        QuestionType = myQ.Mat_Question_Type,
-                        AnswerText = answer?.Answer_Text,
-                        AnswerMemo = answer?.Free_Response_Answer,
-                        Options = GetOptions(myQ.Mat_Question_Id),
-                        Followups = GetFollowupQuestions(myQ.Mat_Question_Id),
-                        Comment = answer?.Comment,
-                        MarkForReview = answer?.Mark_For_Review ?? false
-                    };
-
-                    if (_includeText)
-                    {
-                        question.QuestionText = myQ.Question_Text.Replace("\r\n", "<br/>").Replace("\n", "<br/>").Replace("\r", "<br/> ");
-                        question.ReferenceText = myQ.MATURITY_REFERENCE_TEXT.FirstOrDefault()?.Reference_Text;
-                    }
-
-
-                    // Include the corresponding baseline selection if it exists
-                    var baselineAnswer = baselineAllAnswers
-                        .Where(x => x.Question_Or_Requirement_Id == myQ.Mat_Question_Id)
-                        .FirstOrDefault();
-                    if (baselineAnswer != null)
-                    {
-                        question.BaselineAnswerText = baselineAnswer.Answer_Text;
-                        question.BaselineAnswerMemo = baselineAnswer.Free_Response_Answer;
-                    }
-
-
-                    grouping.Questions.Add(question);
-                }
-
-                // Recurse down to build subgroupings
-                GetSubgroups(grouping, sg.Grouping_Id);
-            }
-        }
-
-
-        /// <summary>
-        /// Get questions that are followups to QUESTIONS
-        /// </summary>
-        /// <param name="allQuestions"></param>
-        /// <param name="parentId"></param>
-        /// <returns></returns>
-        private List<Model.Cis.Question> GetFollowupQuestions(int parentId)
-        {
-            var qList = new List<Model.Cis.Question>();
-
-            var myQuestions = allQuestions.Where(x => x.Parent_Question_Id == parentId && x.Parent_Option_Id == null).ToList();
-
-            foreach (var myQ in myQuestions.OrderBy(s => s.Sequence))
-            {
-                var answer = allAnswers.FirstOrDefault(x => x.Question_Or_Requirement_Id == myQ.Mat_Question_Id);
-
-                var question = new Model.Cis.Question()
-                {
-                    QuestionId = myQ.Mat_Question_Id,
-                    Sequence = myQ.Sequence,
-                    DisplayNumber = myQ.Question_Title,
-                    ParentQuestionId = myQ.Parent_Question_Id,
-                    QuestionType = myQ.Mat_Question_Type,
-                    AnswerText = answer?.Answer_Text,
-                    AnswerMemo = answer?.Free_Response_Answer,
-                    Options = GetOptions(myQ.Mat_Question_Id),
-                    Followups = GetFollowupQuestions(myQ.Mat_Question_Id),
-                    Comment = answer?.Comment,
-                    MarkForReview = answer?.Mark_For_Review ?? false
-                };
-
-                if (_includeText)
-                {
-                    question.QuestionText = myQ.Question_Text.Replace("\r\n", "<br/>").Replace("\n", "<br/>").Replace("\r", "<br/> ");
-                    question.ReferenceText = myQ.MATURITY_REFERENCE_TEXT.FirstOrDefault()?.Reference_Text;
-                }
-
-
-                // Include the corresponding baseline selection if it exists
-                var baselineAnswer = baselineAllAnswers
-                    .Where(x => x.Question_Or_Requirement_Id == myQ.Mat_Question_Id)
-                    .FirstOrDefault();
-                if (baselineAnswer != null)
-                {
-                    question.BaselineAnswerText = baselineAnswer.Answer_Text;
-                    question.BaselineAnswerMemo = baselineAnswer.Free_Response_Answer;
-                }
-
-
-                qList.Add(question);
-
-                var followups = GetFollowupQuestions(myQ.Mat_Question_Id);
-                question.Followups.AddRange(followups);
-            }
-
-            return qList;
-        }
-
-
-        /// <summary>
-        /// Build options for a question.
-        /// </summary>
-        /// <param name="questionId"></param>
-        /// <returns></returns>
-        private List<Option> GetOptions(int questionId)
-        {
-            var list = new List<Option>();
-
-
-            var opts = _context.MATURITY_ANSWER_OPTIONS.Where(x => x.Mat_Question_Id == questionId)
-                .OrderBy(x => x.Answer_Sequence)
-                .ToList();
-
-            foreach (var o in opts)
-            {
-                var option = new Option()
-                {
-                    OptionText = o.Option_Text,
-                    OptionId = o.Mat_Option_Id,
-                    OptionType = o.Mat_Option_Type,
-                    Sequence = o.Answer_Sequence,
-                    HasAnswerText = o.Has_Answer_Text,
-                    Weight = o.Weight
-                };
-
-                var ans = allAnswers.Where(x => x.Question_Or_Requirement_Id == o.Mat_Question_Id
-                    && x.Mat_Option_Id == option.OptionId).FirstOrDefault();
-
-                option.AnswerId = ans?.Answer_Id;
-                option.Selected = ans?.Answer_Text == "S";
-                option.AnswerText = ans?.Free_Response_Answer;
-
-
-                // Include the corresponding baseline selection if it exists
-                var baselineAnswer = baselineAllAnswers
-                    .Where(x => x.Question_Or_Requirement_Id == o.Mat_Question_Id && x.Mat_Option_Id == o.Mat_Option_Id)
-                    .FirstOrDefault();
-                if (baselineAnswer != null)
-                {
-                    option.BaselineSelected = baselineAnswer.Answer_Text == "S";
-                    option.BaselineAnswerText = baselineAnswer.Free_Response_Answer;
-                }
-
-
-
-                // Include questions that are a followup to the OPTION
-                var myQuestions = allQuestions.Where(x => x.Parent_Option_Id == o.Mat_Option_Id).ToList();
-
-                foreach (var myQ in myQuestions.OrderBy(s => s.Sequence))
-                {
-                    var answer = allAnswers.FirstOrDefault(x => x.Question_Or_Requirement_Id == myQ.Mat_Question_Id);
-
-                    var question = new Model.Cis.Question()
-                    {
-                        QuestionId = myQ.Mat_Question_Id,
-                        Sequence = myQ.Sequence,
-                        DisplayNumber = myQ.Question_Title,
-                        ParentQuestionId = myQ.Parent_Question_Id,
-                        ParentOptionId = myQ.Parent_Option_Id,
-                        QuestionType = myQ.Mat_Question_Type,
-                        AnswerText = answer?.Answer_Text,
-                        AnswerMemo = answer?.Free_Response_Answer,
-                        Options = GetOptions(myQ.Mat_Question_Id),
-                        Followups = GetFollowupQuestions(myQ.Mat_Question_Id),
-                        Comment = answer?.Comment,
-                        MarkForReview = answer?.Mark_For_Review ?? false
-                    };
-
-                    if (_includeText)
-                    {
-                        question.QuestionText = myQ.Question_Text.Replace("\r\n", "<br/>").Replace("\n", "<br/>").Replace("\r", "<br/> ");
-                        question.ReferenceText = myQ.MATURITY_REFERENCE_TEXT.FirstOrDefault()?.Reference_Text;
-                    }
-
-
-                    // Include the corresponding baseline selection if it exists
-                    baselineAnswer = baselineAllAnswers
-                        .Where(x => x.Question_Or_Requirement_Id == myQ.Mat_Question_Id)
-                        .FirstOrDefault();
-                    if (baselineAnswer != null)
-                    {
-                        question.BaselineAnswerText = baselineAnswer.Answer_Text;
-                        question.BaselineAnswerMemo = baselineAnswer.Free_Response_Answer;
-                    }
-
-
-                    option.Followups.Add(question);
-                }
-
-                list.Add(option);
-            }
-
-            return list;
-        }
-
+       
 
         /// <summary>
         /// CIS answers are different than normal maturity answers
@@ -616,85 +285,7 @@ namespace CSETWebCore.Business.Maturity
         }
 
 
-        /// <summary>
-        /// Placeholder for the eventual true scoring calculator.
-        /// </summary>
-        /// <returns></returns>
-        public Score CalculateGroupingScore(int sectionId)
-        {
-            if (this.QuestionsModel == null)
-            {
-                LoadStructure(sectionId);
-            }
-
-            if (QuestionsModel.Groupings.FirstOrDefault().Questions != null)
-            {
-                FlattenQuestions(QuestionsModel.Groupings.FirstOrDefault()?.Questions);
-
-                if (allWeights.Any())
-                {
-                    var grouped = allWeights.GroupBy(q => q.QuestionText).Select(r => new GroupedQuestions
-                    {
-                        QuestionText = r.Key,
-                        OptionQuestions = r.ToList()
-                    });
-                    var sumGroupWeights = from g in grouped
-                                          select new RollupOptions
-                                          {
-                                              Type = g.OptionQuestions.FirstOrDefault().Type,
-                                              Weight = g.OptionQuestions.FirstOrDefault().Type != "radio"
-                                                  ? g.OptionQuestions.Sum(x => x.Weight) : g.OptionQuestions.MaxBy(x => x.Weight)?.Weight
-                                          };
-                    var sumAllWeights = (decimal)sumGroupWeights.Sum(x => x.Weight);
-                    var totalGroupWeights = from g in grouped
-                                            select new RollupOptions()
-                                            {
-                                                Type = g.OptionQuestions.FirstOrDefault().Type,
-                                                Weight = g.OptionQuestions.FirstOrDefault().Type != "radio"
-                                                    ? g.OptionQuestions.Where(s => s.Selected).Sum(x => x.Weight)
-                                                    : g.OptionQuestions.FirstOrDefault(s => s.Selected)?.Weight
-                                            };
-                    var sumTotalWeights = (decimal)totalGroupWeights.Sum(x => x?.Weight);
-                    decimal total = sumAllWeights != 0 ? sumTotalWeights / sumAllWeights : 0;
-                    return new Score
-                    {
-                        GroupingScore = (int)Math.Round(total * 100, MidpointRounding.AwayFromZero),
-                        Low = 0,
-                        Median = 0,
-                        High = 0
-                    };
-                }
-            }
-
-            return new Score();
-        }
-
-        private void FlattenQuestions(List<Model.Cis.Question> questions)
-        {
-            foreach (var q in questions)
-            {
-                allWeights.AddRange(q.Options.Select(x => new FlatQuestion
-                {
-                    QuestionText = q.QuestionText,
-                    Weight = x.Weight,
-                    Selected = x.Selected,
-                    Type = x.OptionType
-
-                }).ToList());
-                foreach (var o in q.Options)
-                {
-                    if (o.Followups.Any())
-                    {
-                        FlattenQuestions(o.Followups);
-                    }
-                }
-
-                if (q.Followups.Any())
-                {
-                    FlattenQuestions(q.Followups);
-                }
-            }
-        }
+        
 
 
         /// <summary>
