+++ CSETWebNg/src/app/assessment/questions/maturity-questions/nested/maturity-questions-nested/maturity-questions-nested.component.ts
@@ -97,8 +97,7 @@ export class MaturityQuestionsNestedComponent implements OnInit {
           this.section = response.groupings[0];
           this.scoreObject = response.groupingScore;
           this.sectionScore = this.scoreObject.groupingScore;
-
-          this.baselineScore = 33;
+          this.baselineScore = this.scoreObject.baselineGroupingScore;
         }
 
         this.loaded = true;
@@ -160,6 +159,13 @@ export class MaturityQuestionsNestedComponent implements OnInit {
           backgroundColor: ['#cccccc']
         }]
     };
+
+
+    // Remove the baseline bar if there's no baseline score
+    if (!this.baselineScore) {
+      x.datasets = x.datasets.filter(x => x.label != 'Baseline Score');
+    }
+
 
     let opts = {
       scales: { y: { display: false } },
